+++ packages/angular/src/modules/synergy.module.ts
@@ -37,6 +37,7 @@ import {
   SynSwitchComponent,
   SynTagComponent,
   SynTextareaComponent,
+  SynTooltipComponent,
 } from '../components';
 
 const components = [
@@ -71,6 +72,7 @@ const components = [
   SynSwitchComponent,
   SynTagComponent,
   SynTextareaComponent,
+  SynTooltipComponent,
 ];
 
 @NgModule({