+++ packages/components/scripts/vendorism.js
@@ -21,6 +21,7 @@ import {
   vendorSpinner,
   vendorTag,
   vendorTextarea,
+  vendorTooltip,
   vendorTranslations,
   vendorWebTestRunnerConfig,
 } from './vendorism/custom/index.js';
@@ -49,6 +50,7 @@ export const components = [
   'menu-item',
   'menu-label',
   'badge',
+  'tooltip',
   'progress-bar',
   'progress-ring',
   'alert',
@@ -283,6 +285,7 @@ const config = {
       vendorTextarea,
       vendorTranslations,
       vendorTag,
+      vendorTooltip,
       vendorTranslations,
       vendorWebTestRunnerConfig,
     ],