+++ packages/components/scripts/vendorism/custom/index.js
@@ -12,5 +12,6 @@ export * from './select.vendorism.js';
 export * from './spinner.vendorism.js';
 export * from './tag.vendorism.js';
 export * from './textarea.vendorism.js';
+export * from './tooltip.vendorism.js';
 export * from './translations.vendorism.js';
 export * from './webTestRunner.vendorism.js';