+++ packages/vue/src/index.js
@@ -34,3 +34,4 @@ export { default as SynVueSpinner } from './components/SynVueSpinner.vue';
 export { default as SynVueSwitch } from './components/SynVueSwitch.vue';
 export { default as SynVueTag } from './components/SynVueTag.vue';
 export { default as SynVueTextarea } from './components/SynVueTextarea.vue';
+export { default as SynVueTooltip } from './components/SynVueTooltip.vue';